requirements.txt
pyyaml
argparse
semver
requests
watchdog
